requirements.txt
python-telegram-bot==20.7
aiogram==3.4.1
requests==2.32.3
gunicorn==21.2.0
